+++ hw13_http/entity/page.go
@@ -1,0 +1,9 @@
+package entity
+
+type Page struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+	Host        string `json:"host"`
+	Resource    string `json:"resource"`
+	Method      string `json:"method"`
+}
